requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
python-dotenv
sqlalchemy
asyncpg
alembic
greenlet
pgvector
redis[hiredis]
taskiq
taskiq-redis
taskiq-fastapi
taskiq-dashboard
httpx
tenacity
streamlit
python-docx
python-multipart
pdfplumber
gunicorn